Add skiparray_builder interface.
This significantly reduces the overhead when constructing a
skiparray from presorted, unique bindings. Instead of needing
to find the right point to add the key each time, this just
appends to the end of the skiparray, and appends new nodes
and updates the back/forward pointers as necessary.

Add benchmarks for sequential set with builder (with and without
the ascending key check).

Closes #1.